Remove unused err
The declaration `var err` is unused and it causes some static analysis tools (go
guru) to throw an error on this repository. I'm not sure what ought to be done
here, but consider this a start.

In fact, this ought to be a compile time error. See this issue for the go
runtime https://github.com/golang/go/issues/3059