Change request ID from 8 bit to 32 bit
Currently the request ID is using an 8 bit int. It is possible for the
request ID to be 32 bit. If we were to use 32 bit it would greatly
reduce the risk of sending duplicate request IDs.

When polling a Cisco Nexus devcie via BulkWalk, if a duplicate request
ID is sent then the device sends the response for the previous request.