Set the correct length for length >= 128
This resolves #42

BER length encoding is used to represent the length of the snmp data.
There are 2 ways to do so: short form, and long form. If the length is <
128 then you can use the short form, and the length is respresented by
a single byte, if the length is >= 128 then you need to use the long
form. The most significant bit of the 1st byte identifies the form, 0 =
short, 1 = long. If using the long form then the 7 least significant
bits in the 1st byte show how many bytes are used to represent the
length. i.e `10000010` shows that there are a further 2 bytes that will
represent the length.